Beginning addition of Gulp
Gulp works better for my workflow, I am integrating that. This is the first time I've worked with modifying a gem, there are a few things I don't understand and could probably be done better. For instance inserting the stylesheet and application file into the view, I think I did that part wrong as I had to change which file was being referenced before it would work.

Signed-off-by: Patrick Cauley <patrick@cranesandcaffeine.com>